Retrieve torrent and tracker info with a single RPC.
Embedding a t.multicall inside a d.multicall allows us to send only a single RPC
per torrent instead of 1 RPC to retrieve the torrent data and then one
additional RPC for each of the torrent's trackers.

With ~1600 torrents loaded in rtorrent, this reduces the run time
of (mentor) (from a fresh start) from ~5.9 seconds to ~3.4 seconds).