Fix loading Magnet links with "/" (closes #14)
There was a problem related to overloading the "DEL" key to be able to load both
filenames and magnet links.  Unfortunately, this broke for Magnet links with
tracker addresses containing slash ("/").  This was caused by how Emacs handles
file prompts (see `read-file-name').  In this commit, we remove the mentioned
overloading.

We already have a separate function to load magnet links (by default bound to
the lower case "L" key).  This function is here renamed to
`mentor-download-load-magnet-link-or-url'.

This commit also clarifies the documentation to better reflect this new
behaviour.